requirements.txt
requests==2.32.4
pytz==2025.2
PyYAML==6.0.2
beautifulsoup4==4.13.4